requirements.txt
# Core dependencies
aiogram>=3.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0

# Scheduling
APScheduler>=3.10.0

# Location and time
geopy>=2.3.0
astral>=3.2
pytz>=2023.3

# Data handling
pydantic>=2.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Optional: For better performance
uvloop>=0.17.0; sys_platform != "win32"
